contrib: dut-console: Add .log extension to log file name
This makes it more convenient when uploading the logs on buganizer,
as the mime type gets detected automatically and the log can be
displayed immediately (instead of requiring to download it first).

log-X.log is a little bit redundant, but I find it useful to keep
the log- prefix for all these files.

BUG=none
TEST=dut-console, output file ends in .log.

Change-Id: I3721f22e6f4cf1eff039022ab4342a5965a6796d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2547652
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Auto-Submit: Nicolas Boichat <drinkcat@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>