new_variant: update the end-to-end test for dedede
dedede's gen_fit_image.sh moves the generated blobs to their final
place, so have the end-to-end test place the fake fitimages in the
correct place for commit_fitimage.sh to find them. This also
requires an update to new_variant to know where to look for the
fitimages in check_fit_image_files, and that rippled to all of
the reference boards that use fitimages.

BUG=b:173824076
TEST=`testdata/new_variant_fulltest.sh waddldee && echo PASS`

Cq-Depend: chrome-internal:3421249
Change-Id: I0c1517127cea0de0cd8ba81910d27247824ba3c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2553348
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>