findmissing: Improve finding BUG=/TEST= lines
So far the code creating a CL tried to guess if the fixed patch is from
a stable release branch or from a Chrome OS branch by obtaining the
Change-Id of the patch being fixed. This does not work if the fixed patch
was applied through a merge instead of an individual commit. To fix the
problem, pass an explicit flag to create_change to indicate if the new CL
fixes a problem in Chrome OS or in a stable release branch, and use this
flag to determine if searching for BUG=/TEST= lines in the fixed commit
is warranted.

BUG=b:176447763
TEST=Run tool and verify that TEST/BUG lines are added as expected

Change-Id: Id820442ead6c873162b63c29698dd04d76ee6326
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2606649
Tested-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>