findmissing: Do not push already abandoned changes into Gerrit
The following log message is reported repeatedly by the patches robot.

error: failed to push some refs to 'https://chromium.googlesource.com/chromiumos/third_party/kernel'
ERROR: Failed to create change for kernel_sha 84c82ab8f133 fixed by 5150dd85bdfa

As it turns out, the commit exists in Gerrit with Change-Id
I4a987e8bd1d2a2824f4fc848ab6b82330ff8e7dc and is marked as abandoned for
the target branch. Naturally, Gerrit rejects accepting the patch.

When detecting if a patch already exists in Gerrit, also check if it is
marked as abandoned for the target branch. If so, record it as abandoned
in the database and don't try to push it again.

BUG=b:176923589
TEST=Run robot

Change-Id: Ibaeb85c0649900785f71736e2fa17be0c902bb4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2627626
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>