findmissing: Do not try to apply abandoned changes
For already abandoned changes it is a waste of time to try to apply
a patch when trying to get its status. The only item of interest
is if the patch has been applied to a branch anyway after it was
abandoned. We can do that by looking for the subject line in the
branch commit log.

While this may miss situations where a patch is applied under a
different subject, the potential gain is not worth the cost.

BUG=b:177099511
TEST=Run code on robot

Change-Id: If31e48d96e0eed64b495d9f97b327fccba331ecc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2633103
Tested-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>