findmissing: Don't chdir in commit handler class
It turns out that "git -C <dir>" is just as fast as chdir followed by git
commands. With that in mind, stop changing directories for git accesses
since it just adds complexity for no gain.

With this in place, we can move the git reset methods into the class
(previously that wasn't acceptable for the python syntax police since
the methods did not access 'self').

BUG=b:177099511
TEST=Run robot

Change-Id: I4435815b5734aeb098c858a853f179a9ea9c3820
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2641042
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>