findmissing: getopen: Instantiate commit handlers only once per branch
The getopen tool examines either a stable release branch or a Chrome OS
branch, but not both. That means that we can instantiate the commit
handler(s) for it once and re-use them. There is not need to re-instantiate
the handler(s) for each branch, much less for each SHA.

Move instantiation to once per branch for now. Next step willl be to move
it up even further, to instantiate it once per repository, but I want to
explore if there is a better way to implement this before doing that.

BUG=b:177099511
TEST=Run getopen

Change-Id: Ia377d888f794f023195d877fbe67c121c28defbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2641545
Tested-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>