intel_sync_upstream: Add STAGING_NAME variable
This change adds a variable STAGING_NAME instead of using
STAGING_PREFIX-staging everywhere to refer to the staging
repo. Unfortunately, we have a case where the name of FSP staging repo
deviated from the ones before and this is handled in follow up
changes. In order to make it easier to use any name for the staging
repo, this change adds the STAGING_NAME variable.

BUG=None
TEST=None

Change-Id: I4cd6fa7aa4258a44c2ccfc532eb4ef2b01dfd2a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2641052
Tested-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>