findmissing: Determine base tag dynamically
We use the base tag of a given branch primarily to determine
the number of commits in that branch. So far, that means that
stable release candidates include all commits in the stable release,
not only the commits in the release candidates.

We can work around that by determining the base tag dynamically
and use the most recent tag for release candidates instead of using
the branch base tag. This reduces the number of release candidate patches
significantly and speeds up the code working on release candidates
(specifically the getopen script).

Test show that this patch reduces the 'git log' runtime for release
candidate branches significantly. At the same time, the time needed to
calculate the base tag increases, and the time needed for subject
matches decreases. Overall, the time needed for calculating the base tag,
getting the list of commits in a branch, and searching for subject
matches in that branch is reduced by more than 50% for a typical run
of the 'getopen' tool. This plus the previous patch reduce patches robot
runtime by approximately 15%.

BUG=b:177099511
TEST=Run getopen tool

Change-Id: I8b168210890e5be735609aaa5a71d4d709dc0ad1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2651796
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>