findmissing: Make commit_list a class object of commitHandler
If commit_list is an instance object, its contents have to be
re-calculated for each instance. However, the list of commits
in a repository and branch is static for the runtime of the robot
or a script associated with the robot. Once the commit list was
pulled from git, it will not change anymore.

By making the commit list a class object, we avoid having to
re-calculate its contents. This means that "git log" has to be
executed only once for each branch and repository. Doing this
reduces the runtime of 'git log' for a run of the 'getopen' tool
by more than 50% while not affecting any other operation.

BUG=b:177099511
TEST=Run getopen or robot

Change-Id: Ia92492f0bdd927929cac908671cc8aaa44bd356b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2651795
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>