new_variant: fix query string for Original-Change-Id
Add a "message:" prefix to the query string that searches for the
original change ID, so that gerrit knows to search in the commit
message. It used to work, but that was apparently by accident.

This is the perfect illustration of Hyrum's Law: "all observable
behaviors of your system will be depended on by somebody."

BUG=b:183714280
TEST=with this patch, new_variant.py is able to find the coreboot
upstream CL

Change-Id: Ib4922514742544518ade7db41e70ee4686b83ab9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2787079
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Auto-Submit: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>