merge-kernel: Handle files which were deleted as part of merge
The script was unable to handle files which were deleted as part of a merge
but locally modified. Detect the situation and handle those files.
Add a note to the commit log indicating which locally modified files were
deleted as part of a merge.

BUG=b:159852814
TEST=Perform test merge with deleted file

Change-Id: I7edb119ee72035259921c165f6633c4d5692f752
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2855420
Reviewed-by: William K Lin <wklin@google.com>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>