Moving under chromiumos/test namespace
Inverting the namespace to include it under the go package hierarchy
instead so it's more clearer when using it outside of this src tree and
also matches the proto namespacing.

Also renaming executionserver to testexecserver since this gets
installed into /usr/bin in the chroot and it's non obvious without test
in the bin name.

BUG=None
TEST=emerge lro-server test-exec-server

Cq-Depend: chromium:2913071
Change-Id: I0dfec3a71eb094ca22f71d50265b2adf601d6dfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2910632
Commit-Queue: C Shapiro <shapiroc@chromium.org>
Tested-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>