new_variant: add missing fitimage_cmd
new_variant.py asks the user to generate the fitimage, and offers
an example command. However, the string for this command is missing
from Brya's recipe. Fix that.

BUG=None
TEST=run new_variant.py to create firmware for a new variant of
Brya, and observe that it doesn't crash with an error when it goes
to print the message about generating the fitimage.

Change-Id: If8c02a9694bfb0c797dd1b679e7efac2ce63377d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2934216
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Auto-Submit: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>