Adding DUT Server impl + Tests
Splitting up exec and fetch as two portions to have a working tester
earlier. Note that tests require that we create a wrapper for most ssh
items as they are structs and not interfaces and as such cannot be
mocked.

BUG=None
TEST=None

Change-Id: Ieb779551acf36ded1a88b049e0d0699f1595880e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2946488
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>
Commit-Queue: Jaques Clapauch <jaquesc@google.com>
Tested-by: Jaques Clapauch <jaquesc@google.com>
Auto-Submit: Jaques Clapauch <jaquesc@google.com>