(1 of 2) Moving to src tree from portage
Keeping this in the overlay adds no value since it's not invoked by
portage.

This makes cq coverage rules worse and also makes it harder to share
logic across docker builds.

Initially, this was done to keep the build config close to the ebuild
that actually built the binary (for discoverability, but not worth the
tradeoff).

BUG=None
TEST=./build-dockerimage.sh

Change-Id: I7ba763062a1ab1621e32eacc2e13b41aae57598a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2966018
Tested-by: C Shapiro <shapiroc@chromium.org>
Auto-Submit: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Jaques Clapauch <jaquesc@google.com>