Wire in dutidentity to dutserver
This isn't really setup for testing when it's not being used as a grpc
client, which is a lot of overhead for this type of testing.

Will refactor a bit in coming cls to make non grpc based testing a bit
easier of this server impl and then add testing for this integration
bit.

BUG=b:188712103
TEST=todo

Change-Id: I0836d2052af5bef2c7c1ae8c63300dea4fb440e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3035662
Tested-by: C Shapiro <shapiroc@chromium.org>
Auto-Submit: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>
Commit-Queue: Jaques Clapauch <jaquesc@google.com>