volteer2: change reference for coreboot overlay
Volteer2 needs to use 'volteer' as its reference board for the
coreboot source code and the EC source code, but for the coreboot
overlay, it should be based off the 'volteer2' reference. Update
the reference for the coreboot overlay, but make sure that the
coreboot source and EC source still use volteer.

BUG=b:175250013
TEST=run new_variant_fulltest.sh for each of the supported
reference boards and verify that it passes

Change-Id: I4cc470db08ebef822160592d5d385c6b4a3fc92c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3036841
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>