syz-repro-automation: change logdir loop structure
Instead of leasing a new device if the image id changes and not the
model, I changed the data structure and corresponding functionality we use to store this info to only
lease a new device if the model changes. Also changed error handling to
repro other bugs if an error is encountered and added more descriptive
error messages.

BUG=b:193906330
TEST=go run repro.go

Change-Id: I5c5fe3450791cc12b898dfeec1ba5fd6d174588f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3061026
Reviewed-by: Zubin Mithra <zsm@chromium.org>
Commit-Queue: Grant Hugh <ghugh@chromium.org>
Tested-by: Grant Hugh <ghugh@chromium.org>