syz-repro-automation: bug fixes
Fixed the bugs outlined in comment 15 on the bug associated with this
commit. Most importantly, we fetch the second-latest image for the board
now, error handling is updated so a failure on one bug won't stop others
from running, and finishedbugs file writing permissions are fixed.

BUG=b:193906330
TEST=go run repro.go

Change-Id: I86af558850c0668cf50d227dbfcf06555fe86475
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3067753
Commit-Queue: Grant Hugh <ghugh@chromium.org>
Tested-by: Grant Hugh <ghugh@chromium.org>
Reviewed-by: Zubin Mithra <zsm@chromium.org>