kernel-rebase: genstats: Add short topic statistics
The short statistics don't include topics which are expected to not be
carried along during rebases, topics which are expected to be squashed,
and topics which will never be sent upstream. The resulting chart is
more useful for presentations than the complete topic chart since it
provides a better notion of the real upstream backlog.

BUG=None
TEST=Run genstats.py

Change-Id: If744a6ea96934f78e039419baecfd75e99cd9a29
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3116720
Reviewed-by: Marty Faltesek <mfaltesek@google.com>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>