kernel-rebase: genstats: Generate short topics summary
The short topics summary doesn't include topics which are expected to not
be carried along during rebases, topics which are expected to be squashed,
and topics which will never be sent upstream. The resulting chart is more
useful for presentations than the complete topic chart since it provides
a better notion of the real upstream backlog.

BUG=None
TEST=Run genstats.py

Change-Id: I084fb4c26a78f2ca1753cfba94e911a0ee769f1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3116721
Reviewed-by: Marty Faltesek <mfaltesek@google.com>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>