merge-kernel: Improve handling of conflicts
The merge-kernel script currently relies on "git rerere" functionality
to resolve conflicts. Unfortunately, "git rerere" only considers files
which actually have conflicts. If a conflict resolution involves a file
which did not directly have a conflict, git rerere "forgets" it.

This commit improves the aforementioned issue by using "git stash" to
allow handling files not in conflicting list of a merge. During the
conflict resolving process, after manually "git stash" the files, run
the script with '-a' option to apply that on the top of stash list
(stash@{0}) to the merge. Note that merge-kernel will not modify any
entries in the stash list, it's user's responsiblity to manage them.

BUG=b:191430286
TEST=Create a stash entry, run "merge-kernel -a -b b:197080742
     -r <hash before v5.10.60 merge> v5.10.60" and view the result.

Change-Id: I3220830adfb5e751a70faf46f1ed55d42d61b0d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3128920
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: William K Lin <wklin@google.com>
Tested-by: William K Lin <wklin@google.com>