kernel-rebase: store fixups as patch files
Using commit SHA hashes to identify and cherry-pick fixups is not very
portable due to 2 issues:
* all users need to fetch remotes with fixups to be able to apply them
locally. If fixups are stored in patches, then only the patches
repository and rebase_config.py need to be shared, and the issues in
case of mismatch are easier to track.
* git can use the history of a commit to aid autoresolution of
conflicts, which might yield different results on different machines.
The patch-generating mechanism already addresses this problem by
removing index information from patches.

This patch changes rebase.py and rebase_config.py so that the fixups are
actually stored in patches/fixups/FIXUP_NAME.patch and introduces a
helper function fixup() to aid creation of those files.

BUG=b:199262904
TEST=./rebase.py, then triage()

Change-Id: Ibd65bac3e9bbe0ec4625b8810b321dfa8177138a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3148750
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Kamil Koczurek <koczurek@google.com>
Tested-by: Kamil Koczurek <koczurek@google.com>