F20: Test harness specific metadata
Some harnesses need the ability to pass in harness specific metadata
defined in their proto. For example, gtest needs to know the location of
the binary to be executed. Additional harnesses in the future may have
the same need.

Slightly refactor the driver interface to pass in the full
TestCaseMetadata so that harness drives can access harness metadata if
needed. With this, simplified the TestExecServer by pushing helper
methods to the driver package for generating things like test name to
Id map, etc.

Test harness metadata definitions are intended to remain static and
change infrequently once the harness driver is implemented and in place.

BUG=b:198202240
TEST=Updated and ran unit tests. Ran tests through tast and tauto, ensured proper tests executed and
the results file was correct format with correct test IDs.

Change-Id: I77e97de48936e6b75aac657ec775856611b1bed4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3169953
Tested-by: Jesse McGuire <jessemcguire@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Seewai Fu <seewaifu@google.com>