f20 - remove telemetry from container deps
Its baked into the test image it seems, and autotest is directly calling
for that. It should not be needed.

Container size down to ~1.7gb after this and other changes land

BUG=b:200970609
TEST=build-dockerimage.sh & run_container.py using login_LoginSuccess as
the test on a freshly flashed DUt.

Change-Id: I06cc99d20f18ba80f67b336fc2bb1bc38db07edf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3183599
Tested-by: Derek Beckett <dbeckett@chromium.org>
Auto-Submit: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Jesse McGuire <jessemcguire@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>