cros_unibuild_convert: Drop converting mosys_eventlog setting
This setting is being renamed to manual_eventlog_add instead of
mosys_eventlog, since mosys is dropping the eventlog subcommand.

It's applied on a per-overlay basis anyway; there is no reason to
customize this setting on a per-model basis, so we don't need to
convert it from a USE flag.  Just drop the conversion.

BUG=b:203777035
TEST=dry run for hana

Change-Id: If965b491285c393b3792ae42b52cc58722aacf57
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3244106
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Mengqi Guo <mqg@chromium.org>