new_variant_fulltest: fix overlay path and ebuild
A recent change to the overlay path and the ebuilds caused the
end-to-end tests to fail.
* Update the test scripts to use the correct path and ebuild name
for each supported reference board.
* The non-9999 ebuild also changed, and is no longer a symlink, so
update revbump_ebuild.sh to find the right file now that we can't
use the shortcut that it's the only symlink in the directory.
* Remove chromeos-config-bsp from the explicit package list.

BUG=b:198303126
TEST=Run new_variant_fulltest.sh for each supported reference board

Change-Id: Ia582caa543ff3059064635c9be00918554bf08c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3169677
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>