kernel_decode_stack: Set ARCH in environment for decodecode
The decodecode script is called from the decode_stacktrace.sh script and
it needs to have ARCH set so it knows which architecture to dump the
code in assembly form as. Otherwise the ARCH comes from uname -m, which
doesn't work well when running this for an ARM/ARM64 crash on an x86
machine.

BUG=None
TEST=kernel_decode_stack -b trogdor; see assembly make sense

Change-Id: I902a4a1a766e4beecbc382367b9ee42fbda3dd0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3309762
Tested-by: Stephen Boyd <swboyd@chromium.org>
Auto-Submit: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>
Commit-Queue: Pi-Hsun Shih <pihsun@chromium.org>