update_program_fw: Use TAG instead of branch, and allow it to be customised
Rename the BRANCH var to TAG since it is used for both a branch name
and a gerrit hashtag. Also allow it to be customised so that
it can be unique for testing.

BUG=none
TEST=Rebuild dooly firmware

Change-Id: Ifcba4968e2b117e0e30c091bdd8db8f60bbff55a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3367975
Commit-Queue: Andrew McRae <amcrae@google.com>
Tested-by: Andrew McRae <amcrae@google.com>
Auto-Submit: Andrew McRae <amcrae@google.com>
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Sam McNally <sammc@chromium.org>