findmissing: Drop default reviewer, and copy mailing list if there is no reviewer
So far the default code reviewer was groeck@google.com, which
is a bad idea because an individual can be OOO. Drop the default
reviewer. Copy cros-kernel-codereviews@googlegroups.com instead,
but only if the CL is ChromeOS specific and if there is otherwise
no reviewer to reduce noise.

BUG=b:215375528
TEST=Run robot and confirm that mailing list receives all CLs

Change-Id: I3f2f5175457b3174d2f4721aa748d3cc1330c41a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3405075
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>