kernel-rebase: genlib.py: set locale to en_US
genspreadsheet.py hard-codes the argument separator in formulas to be a
comma (e.g =HYPERLINK("#gid=%d","%s")... in line 93), which isn't
guaranteed to work with all locales. For example, if the user's default is
pl_PL, the spreadsheet will show an error concerning unparseable
formulas, as it expects the separator to be a semicolon.

This patch sets the 'locale' property of a newly created spreadsheet to
'en_US' for consistency.

BUG=b:220018579
TEST=run ./genspreadsheet.py, it should create a valid sheet regardless
of your default locale.

Change-Id: I617fedd1ce24a22a47c57e0c67b10769d2a084de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3469097
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Kamil Koczurek <koczurek@google.com>
Tested-by: Kamil Koczurek <koczurek@google.com>