kernel-rebase: allow to optionally disable rebase-spreadsheet.id
In continuous rebase it's desirable that each time genspreadsheet.py is
run, a completely new spreadsheet is created. This meant that we had to
manually delete rebase-spreadsheet.id every time to ensure that.

This patch modifies genlib.py so that if rebase_filename is set to None
in genspreadsheet.py, the entire process of saving and reading SSID from
a file is skipped, and instead create_spreadsheet(..) is called
directly.

BUG=b:220063000
TEST=set rebase_filename to None in genspreadsheet.py, then run it.
rebase-spreadsheet.id should not be created.

Change-Id: Ib067f3ac1b44f49bda3deb275a40c14258f81307
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3469098
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Kamil Koczurek <koczurek@google.com>
Tested-by: Kamil Koczurek <koczurek@google.com>