cft: impl a cloud-build for cros-test
This is a first phase. I think this is enough to work in a builder, and
if it does I will abstract it further such that other services are built
through this, enabling caching for them, and standardizing the build
process.

BUG=b:228645277
TEST=this, verified the image, and the output

Change-Id: I536c5dca518a87e3081ce244db52d05aadb50d65
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3587331
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>
Commit-Queue: Jaques Clapauch <jaquesc@google.com>