cft: move cros-test docker building from dev/test to dev/src
This will bring it into the same dir (dev/src) as all the other docker builders. Right now this is COPY'ing code over from dev/test/python/lib
Once the builder is fully migrated over, the code from that repo will be
rm'd.

The reason to not move it over 1 in big CL is basically to de-risk breaking the existing builder, and would be a yuge CL to review.

This is CL 1 of many for this.

BUG=b:230875828
TEST=run_cros_test still works, cloudbuild works, all imports in these files work.

Change-Id: I084d7eec279e20dc10dd935b0727d8a995ea2340
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3617126
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>
Tested-by: Derek Beckett <dbeckett@chromium.org>