copybot: Add an option for sign-offs
Add a new option, "add-signed-off-by", which will cause the original
commit message to be amended with a signoff.

This option is typically not desired, as Copybot obviously didn't
author the original commits, and cannot sign a CLA on behalf of the
original author(s).  However, "Sean Paul Bot" expects a sign-off from
the uploader, and therefore provide this option for repos using Sean
Paul Bot.

Enable the option for the Flashrom configuration, who wants option.

BUG=b:232030137
TEST=dry run for flashrom, S-o-b added to commit messages

Change-Id: I1d5879faa171b9889d5c6cf5cf56432ce3b9a831
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3673490
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>