pacman: sample efficiency, global refresh/alignment
the pac utility queries accumulators across multiple devices in order to
establish power measurements for a device. This change modifies the
refresh behavior for the measurements to send the "refresh_g" command,
aligning measurements from all pacs. When binning time-domain data from
the test frameworks, this enables the time-domain data to be time
aligned. Additionally, this initializes the I2C "bus" objects outside of
the critical loop, and uses the auto-incrementing pointer to read all of
the accumulator data in one set (allowing for finer-grained ~100ms)
measurements.

BUG=b:235834092
TEST=running pacman.py, running pac_telemetry_logger

Change-Id: I2b09ebd20875334b4339a6bc1b8e642cc9614ec4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3706067
Tested-by: Parker Holloway <jparkerh@google.com>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Parker Holloway <jparkerh@google.com>
Reviewed-by: Robert Zieba <robertzieba@google.com>