forklift: Handle Change-Id edge case
If there is a Change-Id in the upstream commit, this led to a crash in
_format_commit_message due to found['change-id'] getting popped and
messing up all the indices for other entries following it.

If there is a Change-Id in the upstream commit, we should ignore it and
generate a new Change-Id anyways. So check if the index for the
Change-Id is the greatest metadata index we have found, and only use it
in that case.

Repro case: upstream commit 5b723b12301272ed3c6c99c4ad8b43a520f880ea

BUG=none
TEST=Successfully cherry-pick the failure case identified above

Change-Id: I080496838fe24a4c52f1678737d21521ec9724d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3711281
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Tested-by: Drew Davenport <ddavenport@chromium.org>
Commit-Queue: Drew Davenport <ddavenport@chromium.org>