cros_unibuild_convert: Drop device-tree compatible match generation
There is only one ARM board left to migrate to Unibuild, which is
Hana.

In the case of Hana, not only do we not need
device-tree-compatible-match in our config (there's no more than one
compatible we need to match in this board), but we would have issues
using it.

Many hana firmwares appear to expose compatible strings like so:

   google,hana-rev7
   mediatek,mt8183

(notice the absense of a "google,hana").

Mosys used to support this by having an option for "partial match" for
legacy platforms that required it, that is, "google,hana" could match
"google,hana-rev7".  crosid does not support this.

BUG=b:183045220
TEST=Run with maple DUT attached

Change-Id: I14cef280213dddbf3093506c3eba4bb23697a2d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3752241
Tested-by: Eizan Miyamoto <eizan@chromium.org>
Commit-Queue: Eizan Miyamoto <eizan@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Eizan Miyamoto <eizan@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>