merge-kernel: Add -G <cl-number> switch to force reuse of a CL
-L was previously added to aid the LTS team's effort to re-use CLs,
since the ToT CL matching based on hashtags is ineffective
when the patch number advances, which happens when staging multiple
patch updates on a single CL.

The method employed by -L works when only a single open merge CL
exists on a release branch at a time. Otherwise it fails, with a message
listing all existing CLs. This is normally not an issue for less
busy release branches, but was found to be a problem when multiple test
CLs were opened to debug a CQ issue, and there was a parallel need to
work on the existing CL.

This change adds -G so that the user can directly specify which CL to
reuse when this rare condition arises.

Change-Id: Ic80e58c3539aefd326de3fe9daebc1495a38bfa1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3809031
Commit-Queue: Martin Faltesek <mfaltesek@google.com>
Tested-by: Martin Faltesek <mfaltesek@google.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>