merge-kernel: Remove implicit selection of -L with -T switch
Currently when using -T, which auto locates the LTS branch based on
release number, merge-kernel implicitly adds -L, which locates an
existing CL not based on hashtag but by locating the only open merge
CL for that release branch, based on assumption that there is typically
only a single open merge CL on a less busy release branch at a time.

However, there are cases when -L should not be used, and when a staging
branch is implemented it will not be needed or desired.

This change removes implicit selection of -L when using -T.

Change-Id: I50b461a770792492b45058ed88a54f9b949c98aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3809033
Reviewed-by: William K Lin <wklin@google.com>
Tested-by: Martin Faltesek <mfaltesek@google.com>
Commit-Queue: Martin Faltesek <mfaltesek@google.com>