labtunnel: Refactor btpeers cmd to tunnel to dut and use pos arg count
Pretty much every time we want to tunnel to btpeers we want to tunnel
to the dut just like we do for the wificell cmd. Since the btpeers
cmd does not do that, you end up having to run a labtunnel process
with the dut cmd and another with the btpeers cmd, which is annoying.

This updates the btpeers cmd to always also tunnel to the dut like how
the wificell cmd does. It also removes the "--btpeers" flag in favor
of an optional positional argument, as it is a commonly used option,
and it's easier to add a number after the hostname than it is to
awkwardly say "labtunnel btpeers --btpeers 4 dut_hostname" every time.

BUG=None
TEST=btpeers command tested successfully with examples noted in README
and also confirmed that invalid btpeer_count values were handled as
expected.

Change-Id: Ie8482756589e6a18aab1c1fa6f34d8f81debbf09
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3828801
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Tested-by: Jared Bennett <jaredbennett@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>